+++ src/di/src/Container.php
@@ -54,7 +54,6 @@ class Container implements ContainerInterface
 
     /**
      * Container constructor.
-     * @param Definition\DefinitionSourceInterface $definitionSource
      */
     public function __construct(Definition\DefinitionSourceInterface $definitionSource)
     {